Move test lib to its own folder.
The existence of the 'mobly' folder was causing problems for IntegrationTest.py, which was trying to run 'from mobly import asserts' from the local mobly dir instead of the installed version of mobly.
