+++ tests/lib/mobly_sanity_test_config.json
